removed extra console.logs
